+++ cardano/gateway/src/shared/modules/lucid/dtos/packet/fragments.ts
@@ -28,6 +28,10 @@ export type WithTransferModuleReferenceUtxo = {
   transferModuleReferenceUtxo: UTxO;
 };
 
+export type WithRequiredTransferModuleReferenceUtxo = {
+  transferModuleReferenceUtxo: UTxO;
+};
+
 export type WithMockModuleUtxo = {
   mockModuleUtxo: UTxO;
 };