+++ cosmos/cardano-entrypoint/x/vesseloracle/keeper/query.go
@@ -1,7 +1,0 @@
-package keeper
-
-import (
-	"cardano-entrypoint/x/vesseloracle/types"
-)
-
-var _ types.QueryServer = Keeper{}