+++ proto-types/scripts/codegen.js
@@ -56,7 +56,7 @@ function preserveUnsignedVarintHelpers() {
 }
 
 telescope({
-  protoDirs: ["protos/ibc-go"],
+  protoDirs: ["protos/ibc-go", "../cosmos/vesseloracle-v10/proto"],
   outPath: outPath,
   options: {
     logLevel: 0,
@@ -157,7 +157,7 @@ telescope({
     const index_ts = `
     // Auto-generated, see scripts/codegen.js!
 
-    // Exports we want to provide at the root of the "cosmjs-types" package
+    // Exports we want to provide at the root of the "@cardano-ibc/proto-types" package
 
     export { DeepPartial, Exact } from "./helpers";
     `;