+++ proto-types/src/index.ts
@@ -1,5 +1,5 @@
 // Auto-generated, see scripts/codegen.js!
 
-// Exports we want to provide at the root of the "cosmjs-types" package
+// Exports we want to provide at the root of the "@cardano-ibc/proto-types" package
 
 export { DeepPartial, Exact } from "./helpers";