+++ proto-types/CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- Rename the TypeScript package from `proto-types` to
+  `@cardano-ibc/proto-types` and update its repository metadata and consumers.
+
 ## [0.9.0] - 2023-10-25
 
 - Upgrade to Telescope 1.0. This changes a lot of fields from optional to
@@ -104,7 +109,7 @@ No changelog, sorry. Maybe the diff helps.
 
 - wasmd 0.18 types (cosmwasm/wasm/v1/\*)
 
-[unreleased]: https://github.com/confio/cosmjs-types/compare/v0.9.0...HEAD
+[unreleased]: https://github.com/cardano-foundation/cardano-ibc-incubator/commits/main/proto-types
 [0.9.0]: https://github.com/confio/cosmjs-types/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/confio/cosmjs-types/compare/v0.7.2...v0.8.0
 [0.7.2]: https://github.com/confio/cosmjs-types/compare/v0.7.1...v0.7.2